requirements.txt
absl-py==0.6.1
alabaster==0.7.12
astor==0.7.1
astroid==2.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
bokeh==1.0.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cntk==2.6
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.3
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
gast==0.2.0
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.7
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==17.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mmdnn==0.2.3
nbconvert==5.4.0
nbformat==4.4.0
netron==2.7.9
nltk==3.4
notebook==5.7.4
numpy==1.14.6
numpydoc==0.8.0
opencv-python==3.4.5.20
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==5.4.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
psycopg2==2.8.2
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.0
PyQt5==5.9.2
python-dateutil==2.7.5
pytz==2018.7
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
requests==2.20.1
rope==0.11.0
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
sip==4.19.8
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
sqlparse==0.3.0
statsmodels==0.9.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.3
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.24.1
uuid==1.30
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.10.11
